pyproject.toml
[tool.poetry]
name = "trusted-library-database"
version = "0.1.0"
description = ""
authors = ["Andrey Rukin <rukin897@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
sqlalchemy = "^2.0.23"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
